Add DryRunFileUploader which just generates the JSON file
