+++ slimfast-plugin/src/main/java/com/hubspot/maven/plugins/slimfast/DryRunFileUploader.java
@@ -1,0 +1,11 @@
+package com.hubspot.maven.plugins.slimfast;
+
+import java.nio.file.Path;
+
+public class DryRunFileUploader extends BaseFileUploader {
+
+  @Override
+  protected void doUpload(String bucket, String key, Path path) {
+    // do nothing
+  }
+}
